feat(db) conditional read-before-write for partial update
When performing a partial update, some configurations of
a partial table (e.g. nested records) require a read-before-write
so that only the rest of the record (internally stored as JSON)
is retried and can be correctly merged.

To achieve this, we add some coordination between the schema
library and the generic DAO:

* the schema library detects whether an update of an entity
  necessitates a read-before-write and informs the generic DAO
* the generic DAO performs read-before-write on update only
  if needed, and merges the old and new values with help of
  the schema
* the schema library provides a schema-aware merge function,
  so that records are merged, but array entities are not:
  for example, merging `{a = "foo"}` and `{b = "bar"}` should
  produce `{a = "foo", b = "bar"}`, but merging `{a = {1,2}}`
  and `{a = {4,5,6}, b = "bar"}` should produce
  `{a = {1,2}, b = "bar"}` (and not `a = {1,2,6}`, as a
  generic recursive table-merge function would do)