fix(dao) ensure wrappers return errors with matching types
The old DAO returns errors as `nil, err_t`, and the new DAO
returns as `nil, err, err_t`. This commit ensures the wrappers
return the table as the second argument. Since the table contents
are different and we make no efforts to reconcile this, we change
the `on_error` handler of the Admin API so that this function
can handle both old-style and new-style error tables, deferring
to the new-style handler when such a table is detected.